// shared/types/domain.ts
export type City = 'TP. Hồ Chí Minh' | 'Hà Nội' | 'Đà Nẵng';

export interface Cinema {
  id: string;
  name: string;
  address: string;
  city: City;
  hotline: string;
  auditoriumCount: number;
  hasImax: boolean;
}